Allow unauthenticated GET to /surveys/<survey_id> iff survey_type == 'public'.
@jmwohl
I'm not sure if this is the correct behavior, since I don't see why we
wouldn't allow an unauthenticated user to GET an enumerator_only survey.
After all, the user won't be able to submit to it without
authenticating.

However, this is the behavior specified by the API documentation, so I
implemented it.

Let me know what you think.